Inital Commit of Mark it Done script
This script scans your todoist projects for sections with names matching
those in your done_columns config in the config.yml.  This is a feature
I've requested and this is my little hack to make it a thing.  I
schedule it to run using Cron but I have also consider running it as a
service.